+++ lib/src/features/cipher/data/cipher_action.dart
@@ -1,0 +1,4 @@
+enum CipherAction {
+  encrypt,
+  decrypt,
+}