+++ lib/src/features/cipher/data/cipher_algorithm.dart
@@ -1,0 +1,25 @@
+enum CipherAlgorithm {
+  // * symmetric
+  des,
+  tripleDes,
+  blowfish,
+  twofish,
+  aes,
+  chacha20,
+  serpent,
+
+  // * asymmetric
+  rsa,
+  dsa,
+  ecc,
+
+  // * hash
+  md5,
+  sha1,
+  sha2,
+  sha3,
+  ripeMd160,
+  whirlPool,
+  blake2,
+  argon2,
+}