+++ lib/src/features/cipher/presentation/cipher_algorithm/algorithm_selection_screen.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class AlgorithmSelectionScreen extends StatelessWidget {
+  const AlgorithmSelectionScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(),
+      body: ListView(),
+    );
+  }
+}