+++ lib/src/features/cipher/presentation/cipher_mode_state.g.dart
@@ -1,27 +1,26 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of 'brightness_mode_notifier.dart';
+part of 'cipher_mode_state.dart';
 
 // **************************************************************************
 // RiverpodGenerator
 // **************************************************************************
 
-String _$platformBrightnessHash() =>
-    r'777113a1d652f511b626b1f96ac63c520b1d8437';
+String _$cipherModeStateHash() => r'dc389144fd7a829bf3f009f1921ac915a3c83c03';
 
-/// See also [PlatformBrightness].
-@ProviderFor(PlatformBrightness)
-final platformBrightnessProvider =
-    NotifierProvider<PlatformBrightness, Brightness>.internal(
-  PlatformBrightness.new,
-  name: r'platformBrightnessProvider',
+/// See also [CipherModeState].
+@ProviderFor(CipherModeState)
+final cipherModeStateProvider =
+    NotifierProvider<CipherModeState, CipherMode>.internal(
+  CipherModeState.new,
+  name: r'cipherModeStateProvider',
   debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
       ? null
-      : _$platformBrightnessHash,
+      : _$cipherModeStateHash,
   dependencies: null,
   allTransitiveDependencies: null,
 );
 
-typedef _$PlatformBrightness = Notifier<Brightness>;
+typedef _$CipherModeState = Notifier<CipherMode>;
 // ignore_for_file: type=lint
 // ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package